feat: update step navigation to sync URL query param and stepper state
